Override makeSocket to use a short timeout
Usually, the socket creation uses a 1s timeout. This means that if the logserver is unavailable, the programme can block for up to 1 second trying to make a connection. This is undesirable, as we generally don't want our program to slow down when the log server beocmes unavailable. So instead, we fail fast with a timeout of 0.1s.